animations_101_scene sample is now headerless and auto-registered
