+++ src/Samples/src/samples/animations/animations_101_scene.cpp
@@ -1,5 +1,6 @@
-#include <babylon/samples/animations/animations_101_scene.h>
-
+#include <babylon/babylon_common.h>
+#include <babylon/interfaces/irenderable_scene.h>
+#include <babylon/samples/samples_index.h>
 #include <babylon/animations/animation.h>
 #include <babylon/animations/ianimation_key.h>
 #include <babylon/cameras/arc_rotate_camera.h>
@@ -9,104 +10,119 @@
 #include <babylon/lights/point_light.h>
 #include <babylon/materials/standard_material.h>
 #include <babylon/meshes/mesh.h>
-
 namespace BABYLON {
-namespace Samples {
 
-Animations101Scene::Animations101Scene(ICanvas* iCanvas)
-    : IRenderableScene(iCanvas)
-    , _previousTimestamp{Time::highresTimepointNow()}
-    , _materialBox2{nullptr}
-{
-}
+  class StandardMaterial;
+  using StandardMaterialPtr = std::shared_ptr<StandardMaterial>;
 
-Animations101Scene::~Animations101Scene()
-{
-}
+  namespace Samples {
 
-const char* Animations101Scene::getName()
-{
-  return "Animations 101 Scene";
-}
+    /**
+     * @brief Animations 101 scene. Example demonstrating how to add animations.
+     * @see https://www.babylonjs-playground.com/#QYFDDP#1
+     * @see https://doc.babylonjs.com/babylon101/animations
+     */
+    struct Animations101Scene : public IRenderableScene {
 
-void Animations101Scene::initializeScene(ICanvas* canvas, Scene* scene)
-{
-  // Create a light
-  auto light       = PointLight::New("Omni", Vector3(0.f, 100.f, 100.f), scene);
-  light->intensity = 1.f;
+    public:
+      Animations101Scene(ICanvas* iCanvas)
+        : IRenderableScene(iCanvas)
+        , _previousTimestamp{ Time::highresTimepointNow() }
+        , _materialBox2{ nullptr }
+      {
+      }
 
-  // Create a camera
-  auto camera
-    = ArcRotateCamera::New("Camera", 0.f, 0.8f, 100.f, Vector3::Zero(), scene);
+      ~Animations101Scene() override {} ;
 
-  // Attach the camera to the canvas
-  camera->attachControl(canvas, true);
+      const char* getName() override
+      {
+        return "Animations 101 Scene";
+      }
 
-  // Boxes
-  // - Box 1
-  auto box1          = Mesh::CreateBox("Box1", 10.f, scene);
-  box1->position().x = -20.f;
-  // - Box 2
-  auto box2 = Mesh::CreateBox("Box2", 10.f, scene);
+      void initializeScene(ICanvas* canvas, Scene* scene) override
+      {
+        // Create a light
+        auto light = PointLight::New("Omni", Vector3(0.f, 100.f, 100.f), scene);
+        light->intensity = 1.f;
 
-  auto materialBox          = StandardMaterial::New("texture1", scene);
-  materialBox->diffuseColor = Color3::Green();
-  _materialBox2             = StandardMaterial::New("texture2", scene);
+        // Create a camera
+        auto camera = ArcRotateCamera::New("Camera", 0.f, 0.8f, 100.f, Vector3::Zero(), scene);
 
-  // Applying materials
-  box1->material = materialBox;
-  box2->material = _materialBox2;
+        // Attach the camera to the canvas
+        camera->attachControl(canvas, true);
 
-  // Positioning box
-  box2->position().x = 20.f;
+        // Boxes
+        // - Box 1
+        auto box1 = Mesh::CreateBox("Box1", 10.f, scene);
+        box1->position().x = -20.f;
+        // - Box 2
+        auto box2 = Mesh::CreateBox("Box2", 10.f, scene);
 
-  // Creation of a scaling animation with box 1
-  //-------------------------------------------
-  // Create a scaling animation at 30 FPS
-  auto animationBox = Animation::New("tutoAnimation", "scaling.x", 30,
-                                     Animation::ANIMATIONTYPE_FLOAT(),
-                                     Animation::ANIMATIONLOOPMODE_CYCLE());
+        auto materialBox = StandardMaterial::New("texture1", scene);
+        materialBox->diffuseColor = Color3::Green();
+        _materialBox2 = StandardMaterial::New("texture2", scene);
 
-  // Here we have chosen a loop mode, but you can change to :
-  //  Use previous values and increment it
-  //  (BABYLON.Animation.ANIMATIONLOOPMODE_RELATIVE) Restart from initial value
-  //  (BABYLON.Animation.ANIMATIONLOOPMODE_CYCLE) Keep the final value
-  //  (BABYLON.Animation.ANIMATIONLOOPMODE_CONSTANT)
+        // Applying materials
+        box1->material = materialBox;
+        box2->material = _materialBox2;
 
-  // Animation keys
-  std::vector<IAnimationKey> keysBox1{
-    // At the animation key 0, the value of scaling is "1"
-    IAnimationKey(0, AnimationValue(1.f)),
-    // At the animation key 20, the value of scaling is "0.2"
-    IAnimationKey(20, AnimationValue(0.2f)),
-    // At the animation key 100, the value of scaling is "1"
-    IAnimationKey(100, AnimationValue(1.f)),
-  };
+        // Positioning box
+        box2->position().x = 20.f;
 
-  // Adding keys to the animation object
-  animationBox->setKeys(keysBox1);
+        // Creation of a scaling animation with box 1
+        //-------------------------------------------
+        // Create a scaling animation at 30 FPS
+        auto animationBox
+          = Animation::New("tutoAnimation", "scaling.x", 30, Animation::ANIMATIONTYPE_FLOAT(),
+            Animation::ANIMATIONLOOPMODE_CYCLE());
 
-  // Then add the animation object to box1
-  box1->animations.emplace_back(animationBox);
+        // Here we have chosen a loop mode, but you can change to :
+        //  Use previous values and increment it
+        //  (BABYLON.Animation.ANIMATIONLOOPMODE_RELATIVE) Restart from initial value
+        //  (BABYLON.Animation.ANIMATIONLOOPMODE_CYCLE) Keep the final value
+        //  (BABYLON.Animation.ANIMATIONLOOPMODE_CONSTANT)
 
-  // Finally, launch animations on box1, from key 0 to key 100 with loop
-  // activated
-  scene->beginAnimation(box1, 0.f, 100.f, true);
+        // Animation keys
+        std::vector<IAnimationKey> keysBox1{
+          // At the animation key 0, the value of scaling is "1"
+          IAnimationKey(0, AnimationValue(1.f)),
+          // At the animation key 20, the value of scaling is "0.2"
+          IAnimationKey(20, AnimationValue(0.2f)),
+          // At the animation key 100, the value of scaling is "1"
+          IAnimationKey(100, AnimationValue(1.f)),
+        };
 
-  // Creation of a manual animation with box 2
-  //------------------------------------------
-  scene->registerBeforeRender([this](Scene* /*scene*/, EventState& /*es*/) {
-    auto currentTimestamp = Time::highresTimepointNow();
-    auto diffMillis = Time::fpTimeDiff<size_t, std::milli>(_previousTimestamp,
-                                                           currentTimestamp);
-    if (diffMillis >= 1000ul) {
-      _previousTimestamp = currentTimestamp;
-      // The color is defined at run time with random()
-      _materialBox2->diffuseColor
-        = Color3(Math::random(), Math::random(), Math::random());
-    }
-  });
-}
+        // Adding keys to the animation object
+        animationBox->setKeys(keysBox1);
 
-} // end of namespace Samples
+        // Then add the animation object to box1
+        box1->animations.emplace_back(animationBox);
+
+        // Finally, launch animations on box1, from key 0 to key 100 with loop
+        // activated
+        scene->beginAnimation(box1, 0.f, 100.f, true);
+
+        // Creation of a manual animation with box 2
+        //------------------------------------------
+        scene->registerBeforeRender([this](Scene* /*scene*/, EventState& /*es*/) {
+          auto currentTimestamp = Time::highresTimepointNow();
+          auto diffMillis = Time::fpTimeDiff<size_t, std::milli>(_previousTimestamp, currentTimestamp);
+          if (diffMillis >= 1000ul) {
+            _previousTimestamp = currentTimestamp;
+            // The color is defined at run time with random()
+            _materialBox2->diffuseColor = Color3(Math::random(), Math::random(), Math::random());
+          }
+          });
+      }
+
+    private:
+      high_res_time_point_t _previousTimestamp;
+      StandardMaterialPtr _materialBox2;
+
+    }; // end of struct Animations101Scene
+
+    BABYLON_REGISTER_SAMPLE("Animations", Animations101Scene)
+
+  } // end of namespace Samples
 } // end of namespace BABYLON
+
